requirements.txt
pymongo==3.10.1
youtube-dl==2020.1.24
python-telegram-bot==12.4.2